project/requirements.txt
asgiref==3.8.1
contourpy==1.2.1
cycler==0.12.1
Django==5.0.6
django-filter==24.2
fonttools==4.52.4
kiwisolver==1.4.5
matplotlib==3.9.0
numpy==1.26.4
packaging==24.0
pillow==10.3.0
plotly==5.22.0
psycopg==3.1.19
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.0
tenacity==8.3.0
typing_extensions==4.12.2
tzdata==2024.1
